feat(protocol)!: widen model-turn limits and ordinals to u32 (HC1)
RunLimits.max_model_turns, every turn_ordinal on events, snapshots, and
store rows, the core turn loop bound, the budget meter's counter, and the
CLI's --max-turns were u16, so a supervisor could not request more than
65 535 turns and the loop failed with "exhausted the durable u16 ordinal
space". All of them are u32 now. SQLite INTEGER is 64-bit, so no store
migration is needed.

PROTOCOL_VERSION 17 -> 18. The accepted wire range grew; no field was
renamed or removed, so every version-17 record still decodes. The fixture
harness reads tests/fixtures/v<PROTOCOL_VERSION>/ for the byte-exact
goldens and keeps v17 decode-only under a new historical_fixtures_still_
decode test. Boundary tests pin 65 536 and u32::MAX on the wire and drive
the budget meter past 65 535 without a provider.

BREAKING: a version-17 client rejects a max_model_turns or turn_ordinal
above 65 535; clients and servers must agree on version 18.

Slice: speed-first HC1 (--max-turns width)

+++ crates/qq-core/src/runtime/events.rs
@@ -58,7 +58,7 @@ impl PreparedStaticPrefix {
 pub(crate) enum RuntimeEvent {
     Started,
     Prepared {
-        turn_ordinal: u16,
+        turn_ordinal: u32,
         identity: Option<Arc<RunPromptIdentity>>,
         static_prefix: PreparedStaticPrefix,
         weight: PreparedRequestWeight,
@@ -83,7 +83,7 @@ pub(crate) enum RuntimeEvent {
         text: String,
     },
     AssistantTurnCompleted {
-        turn_ordinal: u16,
+        turn_ordinal: u32,
         message: Message,
         usage: Option<TokenUsage>,
         /// Tool calls requested by this turn, in request order. Carried on the
@@ -144,20 +144,20 @@ pub(crate) enum RuntimeEvent {
     /// turn is prepared.
     SteeringApplied {
         message_id: MessageId,
-        turn_ordinal: u16,
+        turn_ordinal: u32,
     },
     /// An interrupting steer aborted turn `turn_ordinal` in flight. Emitted
     /// after the partial turn (if any text streamed) is committed via
     /// `AssistantTurnCompleted` and before its unfinished calls are settled;
     /// the store marks every call of the turn still open as interrupted.
     Interrupted {
-        turn_ordinal: u16,
+        turn_ordinal: u32,
     },
     /// The provider cut turn `turn_ordinal` at its output token limit. Emitted
     /// after the partial turn is committed via `AssistantTurnCompleted`; the
     /// loop then continues with the next turn. `continuation` is 1-based.
     OutputTruncated {
-        turn_ordinal: u16,
+        turn_ordinal: u32,
         continuation: u16,
     },
     Completed,
@@ -176,7 +176,7 @@ pub(crate) enum RuntimeEvent {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) struct RuntimeToolCall {
     pub(crate) id: ToolCallId,
-    pub(crate) turn_ordinal: u16,
+    pub(crate) turn_ordinal: u32,
     pub(crate) call_ordinal: u16,
     pub(crate) provider_call_id: String,
     pub(crate) name: String,
